# DOCUMENTACION Y SCRPT/logs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "SmarWatt Microservices API",
    "description": "API completa para análisis energético y chatbot inteligente",
    "version": "2.0.0",
    "contact": {
      "name": "SmarWatt Team",
      "email": "dev@smarwatt.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Energy IA API Development"
    },
    {
      "url": "http://localhost:5001",
      "description": "Expert Bot API Development"
    }
  ],
  "paths": {
...
    "/message": {
      "post": {
        "summary": "POST /message",
        "description": "Endpoint que recibe cada mensaje del usuario y orquesta la respuesta.\n    IDÉNTICO AL ORIGINAL - PROHIBIDO CAMBIAR",
        "tags": [
          "Expert Bot API"
        ],
        "responses": {
          "200": {
            "description": "Operación exitosa",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Error en la petición"
          },
          "500": {
            "description": "Error interno del servidor"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "message": {
                    "type": "string",
                    "example": "Mensaje de ejemplo"
                  }
                }
              }
            }
          }
        }
      },
      "options": {
        "summary": "OPTIONS /message",
        "description": "Endpoint que recibe cada mensaje del usuario y orquesta la respuesta.\n    IDÉNTICO AL ORIGINAL - PROHIBIDO CAMBIAR",
        "tags": [
          "Expert Bot API"
        ],
        "responses": {
          "200": {
            "description": "Operación exitosa",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Error en la petición"
          },
          "500": {
            "description": "Error interno del servidor"
          }
        }
      }
    },
...
    "/session/start": {
      "post": {
        "summary": "POST /session/start",
        "description": "Endpoint para iniciar el chat, como espera el frontend.\n    IDÉNTICO AL ORIGINAL - PROHIBIDO CAMBIAR",
        "tags": [
          "Expert Bot API"
        ],
        "responses": {
          "200": {
            "description": "Operación exitosa",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Error en la petición"
          },
          "500": {
            "description": "Error interno del servidor"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "message": {
                    "type": "string",
                    "example": "Mensaje de ejemplo"
                  }
                }
              }
            }
          }
        }
      },
      "options": {
        "summary": "OPTIONS /session/start",
        "description": "Endpoint para iniciar el chat, como espera el frontend.\n    IDÉNTICO AL ORIGINAL - PROHIBIDO CAMBIAR",
        "tags": [
          "Expert Bot API"
        ],
        "responses": {
          "200": {
            "description": "Operación exitosa",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Error en la petición"
          },
          "500": {
            "description": "Error interno del servidor"
          }
        }
      }
    },
    "/new-conversation": {
      "post": {
        "summary": "POST /new-conversation",
        "description": "ENDPOINT EMPRESARIAL: Crear nueva conversación guardando la actual.\n    Funcionalidad para botón \"Nueva Conversación\" del frontend.",
        "tags": [
          "Expert Bot API"
        ],
        "responses": {
          "200": {
            "description": "Operación exitosa",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Error en la petición"
          },
          "500": {
            "description": "Error interno del servidor"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "message": {
                    "type": "string",
                    "example": "Mensaje de ejemplo"
                  }
                }
              }
            }
          }
        }
      },
      "options": {
        "summary": "OPTIONS /new-conversation",
        "description": "ENDPOINT EMPRESARIAL: Crear nueva conversación guardando la actual.\n    Funcionalidad para botón \"Nueva Conversación\" del frontend.",
        "tags": [
          "Expert Bot API"
        ],
        "responses": {
          "200": {
            "description": "Operación exitosa",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Error en la petición"
          },
          "500": {
            "description": "Error interno del servidor"
          }
        }
      }
    },
    "/conversation/history": {
      "get": {
        "summary": "GET /conversation/history",
        "description": "NUEVO ENDPOINT EMPRESARIAL: Recuperar historial de conversaciones del usuario.\n    Funcionalidad solicitada para nivel empresarial.",
        "tags": [
          "Expert Bot API"
        ],
        "responses": {
          "200": {
            "description": "Operación exitosa",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Error en la petición"
          },
          "500": {
            "description": "Error interno del servidor"
          }
        }
      }
    },
    "/conversation/<conversation_id>": {
      "delete": {
        "summary": "DELETE /conversation/<conversation_id>",
        "description": "NUEVO ENDPOINT EMPRESARIAL: Borrar conversación específica.\n    Funcionalidad solicitada para nivel empresarial.",
        "tags": [
          "Expert Bot API"
        ],
        "responses": {
          "200": {
            "description": "Operación exitosa",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Error en la petición"
          },
          "500": {
            "description": "Error interno del servidor"
          }
        }
      },
      "options": {
        "summary": "OPTIONS /conversation/<conversation_id>",
        "description": "NUEVO ENDPOINT EMPRESARIAL: Borrar conversación específica.\n    Funcionalidad solicitada para nivel empresarial.",
        "tags": [
          "Expert Bot API"
        ],
        "responses": {
          "200": {
            "description": "Operación exitosa",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Error en la petición"
          },
          "500": {
            "description": "Error interno del servidor"
          }
        }
      }
    },
    "/conversation/feedback": {
      "post": {
        "summary": "POST /conversation/feedback",
        "description": "NUEVO ENDPOINT EMPRESARIAL: Enviar feedback de conversación.\n    Funcionalidad para aprendizaje automático empresarial.",
        "tags": [
          "Expert Bot API"
        ],
        "responses": {
          "200": {
            "description": "Operación exitosa",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Error en la petición"
          },
          "500": {
            "description": "Error interno del servidor"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "message": {
                    "type": "string",
                    "example": "Mensaje de ejemplo"
                  }
                }
              }
            }
          }
        }
      },
      "options": {
        "summary": "OPTIONS /conversation/feedback",
        "description": "NUEVO ENDPOINT EMPRESARIAL: Enviar feedback de conversación.\n    Funcionalidad para aprendizaje automático empresarial.",
        "tags": [
          "Expert Bot API"
        ],
        "responses": {
          "200": {
            "description": "Operación exitosa",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Error en la petición"
          },
          "500": {
            "description": "Error interno del servidor"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Token JWT de autenticación de Firebase"
      }
    }
  }
}